Fix migration issues for batch -1

+++ src/migrations/20260814_120000_checkout_integrity.ts
@@ -1,6 +1,13 @@
 import { sql, type MigrateDownArgs, type MigrateUpArgs } from '@payloadcms/db-postgres'
 
 /**
+ * NOTE: this migration was never applied to production — a build-time hang forced it to
+ * be removed from `prodMigrations` before it ever ran there. `decrement_stock` and
+ * `orders.idempotency_key` were instead created manually via the Supabase SQL editor on
+ * 2026-08-15. This file has been corrected to match what was actually created (numeric,
+ * not integer — `scarves.stock_quantity` is `numeric`), so that running it later against
+ * production is a safe no-op instead of adding a conflicting `decrement_stock` overload.
+ *
  * Checkout integrity: an atomic stock-decrement function used by
  * `POST /api/store/checkout` (one `UPDATE ... WHERE stock_quantity >= qty RETURNING`
  * per line, inside the order-creation transaction — no read-then-write race), plus the
@@ -18,10 +25,10 @@ export async function up({ db }: MigrateUpArgs): Promise<void> {
   `)
 
   await db.execute(sql`
-    CREATE OR REPLACE FUNCTION decrement_stock(p_id integer, p_qty integer)
-    RETURNS integer AS $$
+    CREATE OR REPLACE FUNCTION decrement_stock(p_id integer, p_qty numeric)
+    RETURNS numeric AS $$
     DECLARE
-      v_new_stock integer;
+      v_new_stock numeric;
     BEGIN
       UPDATE "scarves"
       SET "stock_quantity" = "stock_quantity" - p_qty
@@ -35,7 +42,7 @@ export async function up({ db }: MigrateUpArgs): Promise<void> {
 }
 
 export async function down({ db }: MigrateDownArgs): Promise<void> {
-  await db.execute(sql`DROP FUNCTION IF EXISTS decrement_stock(integer, integer);`)
+  await db.execute(sql`DROP FUNCTION IF EXISTS decrement_stock(integer, numeric);`)
   await db.execute(sql`DROP INDEX IF EXISTS "orders_idempotency_key_idx";`)
   await db.execute(sql`ALTER TABLE "orders" DROP COLUMN IF EXISTS "idempotency_key";`)
 }
